pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-based-search-drones"
version = "0.1.0"
authors = [
    {name = "22930", email = "22930@nsu.ru"},
]
description = "Vision-based drone search system"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.10.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "onnxruntime>=1.16.0",
    "GitPython>=3.1.0",
    "dvc>=3.0.0",
    "pymavlink",
    "geopy",
    "pre-commit>=3.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
]

[project.scripts]
vision-train = "src.cli:train"
vision-infer = "src.cli:infer"